--- src/__mocks__/nativeModules.js
@@ -1,5 +1,0 @@
-jest.mock('react-native-fs', () => {
-  return {
-    mkdir: jest.fn(),
-  };
-});
