+++ src/utils/SerialParsingHelpers.spec.tsx
@@ -1,4 +1,6 @@
-import { getAlarmValues } from './SerialParser';
+import { getAlarmValues } from './SerialParsingHelpers';
+
+jest.useFakeTimers();
 
 test('check System Reset value', () => {
   let testPacket: number[] = new Array<number>(49).fill(0);
@@ -8,4 +10,8 @@ test('check System Reset value', () => {
   expect(alarms[0]).toBe(
     'System Reset - Please Recheck FiO2/PEEP/Minute Ventilation Settings',
   );
+});
+
+afterAll(() => {
+  jest.runAllTimers();
 });
